#12-1 File Name Refactoring
Components - BottomTab.js Header.js Select.js
Screen - HomePage.js LoginPage.js Main.js MyPage.js
