+++ App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect } from 'react';
 import {
   View,
   Button,
@@ -6,16 +6,16 @@ import {
   Text,
   StyleSheet,
   TouchableOpacity,
-} from "react-native";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { Fontisto } from "@expo/vector-icons";
-import { StatusBar } from "expo-status-bar";
+} from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { Fontisto } from '@expo/vector-icons';
+import { StatusBar } from 'expo-status-bar';
 
-import HomeHeader from "./src/Components/HomeHeader";
-import Login from "./src/Screen/Login";
-import MainHome from "./src/Screen/MainHome";
+import HomeHeader from './src/Components/Header';
+import LoginPage from './src/Screen/LoginPage';
+import Main from './src/Screen/Main';
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
@@ -26,14 +26,14 @@ export default function App() {
       <Stack.Navigator initialRouteName="Login">
         <Stack.Screen
           name="Login"
-          component={Login}
+          component={LoginPage}
           options={{
             headerShown: false, // 상단에 흰색 바가 생기고 HOME이라는 글씨가 쓰여있는데 그거 안 보이게 하는 속성
           }}
         />
         <Stack.Screen
           name="Main_Home"
-          component={MainHome}
+          component={Main}
           options={{
             headerShown: false,
           }}
