+++ src/Components/Header.js
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from "react";
-import { Text, View, TouchableOpacity, StyleSheet } from "react-native";
-import { Fontisto, MaterialCommunityIcons } from "@expo/vector-icons";
+import React, { useState, useEffect } from 'react';
+import { Text, View, TouchableOpacity, StyleSheet } from 'react-native';
+import { Fontisto, MaterialCommunityIcons } from '@expo/vector-icons';
 
 // Fontisto 부분 size 수정해줘야함 24나 32로 해놓으면 휴대폰 스크린 크기에 따라 크고 작게 보일 수 있으므로
 
-function HomeHeader() {
+function Header() {
   const notice = () => {}; // 종 모양 눌렀을 때 발생할 이벤트
 
   const user = () => {}; // 사람 모양 눌렀을 때 발생할 이벤트
@@ -12,9 +12,9 @@ function HomeHeader() {
     <View
       style={{
         flex: 1,
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "flex-end",
+        flexDirection: 'row',
+        alignItems: 'center',
+        justifyContent: 'flex-end',
         marginRight: 20,
       }}
     >
@@ -34,4 +34,4 @@ function HomeHeader() {
 
 const style = StyleSheet.create({});
 
-export default HomeHeader;
+export default Header;
