+++ src/Screen/HomePage.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect } from 'react';
 import {
   View,
   Button,
@@ -6,15 +6,15 @@ import {
   Text,
   StyleSheet,
   TouchableOpacity,
-} from "react-native";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { Fontisto } from "@expo/vector-icons";
-import { StatusBar } from "expo-status-bar";
+} from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { Fontisto } from '@expo/vector-icons';
+import { StatusBar } from 'expo-status-bar';
 
-import BottomTab from "../Components/BottomTab";
-import HomeHeader from "../Components/HomeHeader";
+import BottomTab from '../Components/BottomTab';
+import HomeHeader from '../Components/Header';
 
 function Home({ navigation }) {
   return (
@@ -37,15 +37,15 @@ function Home({ navigation }) {
 
 const styles = StyleSheet.create({
   homeMenu: {
-    backgroundColor: "#CCCCCC",
-    borderRadius: "15%",
+    backgroundColor: '#CCCCCC',
+    borderRadius: '15%',
     paddingHorizontal: 10,
     paddingVertical: 8,
-    width: "80%",
+    width: '80%',
   },
   homeMenu_detail: {
-    flexDirection: "row",
-    justifyContent: "space-between",
+    flexDirection: 'row',
+    justifyContent: 'space-between',
   },
   homeMenu_Text: {
     fontSize: 18,
